tests/wpt/web-platform-tests/domxpath/evaluator-constructor.html

<!doctype html>
<meta charset=utf-8>
<title>XPathEvaluator constructor</title>
<link rel=help href="http://wiki.whatwg.org/wiki/DOM_XPath">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
  var x = new XPathEvaluator();
  assert_true(x instanceof XPathEvaluator);
}, "Constructor with 'new'");
test(function() {
  assert_throws(new TypeError(), "var x = XPathEvaluator()");
}, "Constructor without 'new'");
</script>
